added div card value tooltips, added drop rates into calculation

+++ DivMaps/Models/MapViewModel.cs
@@ -3,13 +3,13 @@ namespace DivMaps.Models
     public class MapViewModel
     {
         public string Name { get; set; }
-        public int Value { get; set; }
+        public double Value { get; set; }
         public List<DivCardViewModel> ValueSources { get; set; }
         public MapViewModel()
         {
 
         }
-        public MapViewModel(string name, int value, List<DivCardViewModel> valueSources)
+        public MapViewModel(string name, double value, List<DivCardViewModel> valueSources)
         {
             Name = name;
             Value = value;
